pyproject.toml
[tool.poetry]
name = "postagram-backend"
version = "0.1.0"
description = ""
authors = ["msgtv <linar_97@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.8.0"
django-cors-headers = "^4.3.1"
drf-nested-routers = "^0.94.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
